@@ -1,58 +1,48 @@
-import React from 'react';
-import './App.css';
-import Header from '../Header/Header';
-import Footer from '../Footer/Footer';
-import Login from '../Login/Login';
+import React from "react";
+import Header from "../Header/Header";
+import Footer from "../Footer/Footer";
+import Login from "../Login/Login";
+import CourseList from "../CourseList/CourseList";
 import Notifications from "../Notifications/Notifications";
-import PropTypes from 'prop-types';
+import "./App.css";
+import PropTypes from "prop-types";
+import { getLatestNotification } from "../utils/utils";
 
 class App extends React.Component {
-	// constructor(props) {
-	// 	super(props);
-	// 	this.handleKeydown = this.handleKeydown.bind(this);
-	// }
+  listCourses = [
+    { id: 1, name: "ES6", credit: 60 },
+    { id: 2, name: "Webpack", credit: 20 },
+    { id: 3, name: "React", credit: 40 },
+  ];
 
-	// componentDidMount() {
-	// 	window.addEventListener("keydown", this.handleKeydown)
-	// }
+  listNotifications = [
+    { id: 1, type: "default", value: "New course available" },
+    { id: 2, type: "urgent", value: "New resume available" },
+    { id: 3, type: "urgent", html: getLatestNotification() },
+  ];
 
-	// componentWillUnmount() {
-	// 	window.removeEventListener("keydown", this.handleKeydown)
-	// }
-
-	// handleKeydown(event) {
-	// 	const { logOut } = this.props;
-	// 	if (event.ctrlKey && event.key === 'h') {
-	// 		alert('Logging you out');
-	// 		logOut();
-	// 	}
-	// }
-
-	render() {
-		return (
-			<>
-				<div className='root-notifications'>
-					<Notifications />
-				</div>
-				<div className='App'>
-					<Header />
-					<Login />
-					<Footer />
-				</div>
-			</>
-		);
-	}
-
+  render() {
+    return (
+      <React.Fragment>
+        <div className="App">
+          <div className="heading-section">
+            <Notifications listNotifications={this.listNotifications} />
+            <Header />
+          </div>
+          {this.props.isLoggedIn ? <CourseList listCourses={this.listCourses} /> : <Login />}
+          <Footer />
+        </div>
+      </React.Fragment>
+    );
+  }
 }
 
+App.defaultProps = {
+  isLoggedIn: false,
+};
 
 App.propTypes = {
-	logOut: PropTypes.func,
+  isLoggedIn: PropTypes.bool,
 };
-
-App.defaultProps = {
-	logOut: () => { },
-};
-
 
 export default App;
